+++ src/modify/cvolton.betterinfo/CustomCreatorLayer.cpp
@@ -1,9 +1,10 @@
 #include <Geode/Geode.hpp>
 #include "../../SwelvyBG.hpp"
-#include "../../Hooks/Hooker.hpp" 
+#include <external-hook-api/HookClass.hpp>
 using namespace geode::prelude;
 
-Viper_Hookclass_Scene(cvolton_betterinfo_CustomCreatorLayer,"cvolton.betterinfo/CustomCreatorLayer") {
+HookClass_Scene(cvolton_betterinfo_CustomCreatorLayer,"cvolton.betterinfo/CustomCreatorLayer") {
+    if (Mod::get()->getSettingValue<bool>("external-mods")) {
          if (auto bg = _This->getChildByID("cvolton.betterinfo/background")) {
             bg->setVisible(false);
             SwelvyBG* swelvyBG = SwelvyBG::create();
@@ -11,4 +12,5 @@ Viper_Hookclass_Scene(cvolton_betterinfo_CustomCreatorLayer,"cvolton.betterinfo/
             swelvyBG->setID("swelvy-background");
             _This->addChild(swelvyBG);
         }
+    }
 }