+++ src/modify/cvolton.betterinfo/LevelSearchViewLayer.cpp
@@ -1,14 +1,16 @@
 #include <Geode/Geode.hpp>
 #include "../../SwelvyBG.hpp"
-#include "../../Hooks/Hooker.hpp" 
+#include <external-hook-api/HookClass.hpp> 
 
 using namespace geode::prelude;
-Viper_Hookclass_Scene(cvolton_betterinfo_LevelSearchViewLayer,"cvolton.betterinfo/LevelSearchViewLayer") {
- if (auto bg = _This->getChildByID("cvolton.betterinfo/background")) {
+HookClass_Scene(cvolton_betterinfo_LevelSearchViewLayer,"cvolton.betterinfo/LevelSearchViewLayer") {
+    if (Mod::get()->getSettingValue<bool>("external-mods")) {
+        if (auto bg = _This->getChildByID("cvolton.betterinfo/background")) {
             bg->setVisible(false);
         }
         SwelvyBG* swelvyBG = SwelvyBG::create();
         swelvyBG->setZOrder(-1);
         swelvyBG->setID("swelvy-background");
         _This->addChild(swelvyBG);
+    }
 }