+++ src/modify/cvolton.betterinfo/RewardGroupLayer.cpp
@@ -1,8 +1,9 @@
 #include <Geode/Geode.hpp>
 #include "../../SwelvyBG.hpp"
-#include "../../Hooks/Hooker.hpp" 
+#include <external-hook-api/HookClass.hpp> 
 using namespace geode::prelude;
-Viper_Hookclass(RewardGroupLayer) {
+HookClass(RewardGroupLayer) {
+    if (Mod::get()->getSettingValue<bool>("external-mods")) {
         if (auto bg = this->getChildByID("cvolton.betterinfo/background")) {
             bg->setVisible(false);
         }
@@ -10,4 +11,5 @@ Viper_Hookclass(RewardGroupLayer) {
         swelvyBG->setZOrder(-1);
         swelvyBG->setID("swelvy-background");
         this->addChild(swelvyBG);
+    }
 }