+++ src/modify/geode.texture-loader/PackSelectLayer.cpp
@@ -1,10 +1,11 @@
 #include <Geode/Geode.hpp>
 #include "../../SwelvyBG.hpp"
-#include "../../Hooks/Hooker.hpp"
+#include <external-hook-api/HookClass.hpp>
 
 using namespace geode::prelude;
 
-Viper_Hookclass(PackSelectLayer) {
+HookClass(PackSelectLayer) {
+    if (Mod::get()->getSettingValue<bool>("external-mods")) {
         if (auto bg = this->getChildByID("background")) {
             bg->setVisible(false);
         }
@@ -13,4 +14,5 @@ Viper_Hookclass(PackSelectLayer) {
         swelvyBG->setZOrder(-1);
         swelvyBG->setID("swelvy-background");
         this->addChild(swelvyBG);
+    }
 }