+++ src/modify/hiimjustin000.integrated_demonlist/IDListLayer.cpp
@@ -1,9 +1,9 @@
 #include <Geode/Geode.hpp>
 #include "../../SwelvyBG.hpp"
-#include "../../Hooks/Hooker.hpp" 
+#include <external-hook-api/HookClass.hpp> 
 
 using namespace geode::prelude;
-Viper_Hookclass(IDListLayer) {
+HookClass(IDListLayer) {
         if (auto bg = this->getChildByType<CCSprite>(0)) {
             bg->setVisible(false);
         }