+++ src/modify/limegradient.betterachievements/BetterAchievementLayer.cpp
@@ -1,9 +1,10 @@
 #include <Geode/Geode.hpp>
 #include "../../SwelvyBG.hpp"
-#include "../../Hooks/Hooker.hpp" 
+#include <external-hook-api/HookClass.hpp> 
 
 using namespace geode::prelude;
-Viper_Hookclass(BetterAchievementLayer) {
+HookClass(BetterAchievementLayer) {
+    if (Mod::get()->getSettingValue<bool>("external-mods")) {
         if (auto bg = this->getChildByID("content-background")) {
             bg->setVisible(false);
         }
@@ -12,6 +13,7 @@ Viper_Hookclass(BetterAchievementLayer) {
         swelvyBG->setZOrder(-1);
         swelvyBG->setID("swelvy-background");
         this->addChild(swelvyBG);
+    }
 }