+++ src/modify/minemaker0430.gddp_integration/DPListLayer.cpp
@@ -1,9 +1,10 @@
 #include <Geode/Geode.hpp>
 #include "../../SwelvyBG.hpp"
-#include "../../Hooks/Hooker.hpp" 
+#include <external-hook-api/HookClass.hpp> 
 
 using namespace geode::prelude;
-Viper_Hookclass(DPListLayer) {
+HookClass(DPListLayer) {
+    if (Mod::get()->getSettingValue<bool>("external-mods")) {
         if (auto bg = this->getChildByID("bg")) {
             bg->setVisible(false);
         }
@@ -12,4 +13,5 @@ Viper_Hookclass(DPListLayer) {
         swelvyBG->setZOrder(-99999);
         swelvyBG->setID("swelvy-background");
         this->addChild(swelvyBG);
+    }
 }